Fixed a crash on the #enabled? method
A user reported a crash when you perform
`cfg.enabled?("something.another")` and `something` is a key that
doesn't exist. The fix for this has been to implement a proper general
algorithm for nested cases.

Moreover, I've updated the Gemfile.lock file, since it was still
accounting for the 1.0.0 version...

Fixes #3

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>