railtie: added a fix for Rails 6.x
The `Module#parent_name` method is being deprecated and it will be removed in
Rails 6.1. We should be using `Module#module_parent_name` instead, which is
already available in Rails 6.0.

From now on we will use the new method for Rails 6 onwards, and the old method
for previous versions of Rails.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>